Fix file change detection in dotc & dotr scripts
`>&` redirects both stdout and stderr, here we *need* `stdout` for the
test below but don't care about `stderr`: when
`dist-bootstrapped/target/pack/VERSION` doesn't exists, users don't
should't see the error from find.

Also testing for the existance of `$target` to rebuild dotty is obviously
a typo, we want to rebuild `$version` is not on disk.