Fix #3597 - Don't check inline for params after pickling
There was a discrepancy in that value parameters had a Deferred flag set
before pickling but not after unpickling. This triggered a check that the
(missing) rhs of an inline parameter was a constant. This commit changes
the condition of the test to also exclude parameters. It also aligns
frontend and unpickler in that the frontend will no longer mark term
parameters as Deferred.