Inline by-name parameters directly
Do not bind them to an auxiliary method. This has the advantage that
we pass the actual quoted arguments to a macro implementation instead
of passing proxies.

To avoid code duplication, inline methods using an argument more than once
should create a local auxiliary method instead. This is done in `trace.scala`
and `Stats.scala`.