Don't require implicits to make type refs persist stages
It turns out that we can always make type references cross-stage
persistent because we can create a `quoted.Type[T]` from any type `T`.
So implicit search is not needed, after all.

This means we don't need anymore the infrastructure of allowing implicit
searches in macro transforms. But we don't remove it because it might become
useful later.