Fix #1905: Detect case where bridge would clash with the member
With value classes it is possible that a bridge is needed because the semi erased
types of a member and its overridden method differ, yet the fully erased types
of the member and the bridge are the same. With this commit we flag this situation
as an error. Before it generated a ClassCastException at runtime instead.