Don't make superAccessors private
The previous logic made a super accessor private if it is put in a class (on behalf
of a nested trait). This caused compilation from Tasty to fail, since at the point
of Tasty generation the RHS of the accessor has not yet been generated (it will be
generated later in resolveSuper), and therefore Unpickler concluded that the method
is private and deferred, which is illegal.

Making it private is anyway pointless, since the super accessor is referred to from
an inner trait (that's why it was generated in the first place!) and therefore will
be made non-private later on.