Fix #3533: Fix parsing of raw types appearing in generic position
Only the ClassfileParser is fixed (because we can just ask users to fix
their Java source files to not use raw types to get them to compile with
Dotty). This means that the added Java testcase does not pass Dotty's
Java source parser, we handle this by putting "JAVA_ONLY" in the filename and
skipping files that contain this string from going through Dotty in our
testing framework.