Fix #3145: treat sealed classes and traits as inhabited
Sealed classes and traits are effectively inhabited in
practice. By treating them as inhabited by default,
we avoid spurious warnings that ignore anonymous classes.

When we have a mechanism to track anonymous children or
disallow anonymous children for sealed classes, we
could revert this change to make the check stronger.