Refactor ClassSymbol#tree
Use less state: just a single field that can be either a tree or a tree provider
This should make it easy to integrate sourfe files in IDE.

The refactoring required a major refactoring in ReadTastyTreesFromClasses, which
was a mess before. Hopefully it's not right. There are still "spurious" warnings
coming from classes loaded twice. These will require a refactoring of FromTastyTest
to avoid. They do not happen if the FromTasty frontend is invoked stand-alone.