JavaParsers: fix call to annotation methods
Previously, for tests/run/i2760/Fork.java, the parser output was:

    abstract class Fork(val value: Int = ???,  val warmups: Int = ???)
    extends ...

Since Fork is JavaDefined, calls to `value` were translated into
field calls, but `value` is a method, not a field, so i2760 crashed at
runtime with NoSuchFieldError. We now generate instead:

    abstract class Fork private[this](x$1: _root_.scala.Unit)
    extends ... {
       def <init>(value: Int = ???, warmups: Int = ???)

       def value(): Int = 1
       def warmups(): Int = 1
    }

This way we get both named parameters for the constructor and real
method calls.