Tighten matchSymbol again
Revert to the original definition. For a while this was changed
so that a tree would match if its source symbol matched the required
symbol but this yields to many trees (e.g. primary constructor of
a class in addition to the class itself). (reverted from commit 349d79bdbcab46f5021d06100a74b57d00834a83)