Move hashing to separate object
This is a first step of a larger refactoring, where hashing and equality need
to have algorithms that detect that isomorphic BindingTypes are equal.

Without this, there is no good way to deal with recursive higher-kinded types,
as we cannot detect that a type has already been added to a constraint or that
a subtype test was already performed.