Access check secondary constructor parameters
We handle a supercall of a secondary constructor specially,
preventing accesses to members of the enclosing class. We need to
do the same for parameters of such constructors.

We now use two different mechanisms for that: superCallContexts for the
super call and checkRefsLegal for the parameters. We should experiment
at some point with doing the supercall checks also with chekRefsLegal.