Follow GADT bounds when computing members
When computing the member denotation of a selection of a TypeRef `T`,
if the normal scheme fails and `T` has GADT bounds, compute the member
in the upper bound instead.

This is needed to make the opaque-probability test work.
Add this test, as well as some others coming from SIP 35.