Consider GADT bounds for findMember and underlying
The `findMember` and `underlying` methods on `TermRef` now take GADT bounds into account.
This replaces the previous special case in `secectionType`. The previous case did always
pass -Ycheck.

Currently, the treatment is restricted to TypeRefs of opaque types. We should extend
that to all GADT-bound TypeRefs. The question is how this will affect performance.