Add Any-kinded types
Allow a form of type polymorphism by adding an AnyKind upper bound.

Types of any kind are subtypes of `AnyKind`. The "any-kinded types" are `AnyKind` itself, and all type
parameters and abstract types having it as an upper bound.

Any-kinded types cannot be used as types of values, nor can they be applied to type parameters. About the
only thing one can do with them is use them as parameters for implicit searches.