Harden tryReduce
When compiling kind-incorrect code in neg/anykind3.scala, substParams fails with
an index-out-bounds exception. It is called from tryReduce via instantiate. We
know there are several scenarios where this can happen - Sandro has detected
a couple of others. This commit avoids the problem by aborting the reduce
if there is an index out of bounds.