[Proposal] Common Declarations
This is a proposal to add `common` declarations to the language. These declarations are an important stepping stone
to full Rust-like type classes. Given what's described here it would be a relatively small step to do the full typeclass encoding.
Without the full encoding, the usefulness of `common` definitions is still there but not as great. So it's dubious
whether this would pay its cost as a separate feature. Nevertheless, it's nicer to describe this feature in isolation,
since we achieve better conceptual modularity that way.