Fix denotation caching in NamedType
We cannot rely on the validity period cached in checkedPeriod to avoid
calling `Denotation#current` because `SymDenotation#installAfter` can
change the validity period of a denotation.

This change appears to have no effect on performance;

This change unmasked a bug in `isMethodWithExtension`: we need to be
careful to not pick up methods added in the class in phases after
ExtensionMethods was run.