Fix renaming of synthetic positions in IDE
While working on the IDE tests, I noticed that the Dotty IDE didn't
behave as I expected in the following scenario:

  // Rename `Foo` to `Bar`
  class Foo { new Foo }

The Dotty IDE would consider that 3 places need changing, instead of
only 2:

 - The `Ident(Foo)` in the `TypeDef`
 - The `Ident(Foo)` in the constructor
 - The `Ident(Foo)` in `Select(new Foo, <init>)`

The third tree, however is synthetic: it doesn't need to be changed in
the editor.

Trees whose positions are not derived from the source are now excluded
when doing a rename.