Fuse the FirstTransform and CheckStatic group
This requires moving the code that makes sure a module class appears
after its companion class before FirstTransform. We move it to
PostTyper, which means that this ordering now becomes a requirement of
TASTY, hence we bump TASTY to 6.0 and add a conformance check to the
unpickler.

tests/pos/i1812b.scala and tests/patmat/i3645.scala had to be changed
because they contain forward references in the right hand side of a type
definition. We missed those references before because FirstTransform
replace them by bare TypeTrees.