Avoid false companions in scalacLinkedClass
`scalacLinkedClass` should not pick up a companion defined in a
different run than the current class. We can do this by checking for
absent denotations, because in Namer#createCompanionLinks we explicitly
get rid of false companions by creating symbols with absent denotations.

This avoids triggering the assertion added in TreeUnpickler in the
previous commit when running with -Ytest-pickler, and this seems like a
good idea in general.