Optimize miniphase handling of trees representing types
In FirstTransform, a tree representing a type is replaced by a bare
TypeTree, but because of the way miniphases work, its children are first
processed by all the mini-phases in the same group as FirstTransform,
even though this isn't necessary. To avoid this, we now have an explicit
method in MegaPhase to deal with trees representing types:
`transformTpt` which does not recursively transform its children.
This method also replaces `transformTypeTree` since all trees
representing types are `TypeTree` after the `FirstTransform` group.