Restrict auto-tupling more
This was initially an attempt to drop auto-tupling altogether. But it became quickly annoying. The
main problem is that it's unnatural to add another pair of parentheses to infix operations. Compare:

    (x, y) == z
    z == ((x, y))    // !yuck

Same thing for cons:

    ((1, x)) :: ((2, y)) :: xs   // double yuck!

So at the very least we have to allow auto-tupling for arguments of infix operators. I went a little bit
further and also allowed auto-tupling if the expected type is already a (some kind of) product or tuple of the right
arity. The result is this commit.

The language import `noAutoTupling` has been removed. The previous auto-tupling is still reported
under Scala 2 mode, and there is -rewrite support to add the missing (...) automatically.