Avoid infix operations with more than one argument on the right. The idiom
    x op (y, z)

will in the future be interpreted as taking a tuple `(y, z)`. So all
infix operations with more than one parameter have to be rewritten to
method calls. This was for the most part done using an automatic rewrite
(introduced in one of the next commits). Most of the tests were not re-formatted
afterwards so one sees the traces of the rewrite. E.g. the reqrite would yield

    x .op (y, z)

instead of the more idiomatic

    x.op(y. z)