Don't emit forwarder in mirror class for bridge methods
When overriding a (concrete) method in an object with a varying
signature, make sure only one static forwarder method is emitted in the
mirror class, not an additional one with the signature of the bridge.

    class A { def f: Object = null }
    object B extends A { override def f: String = "b" }

`typeOf[B.type].membersBasedOnFlags(BRIDGE, METHOD)` after erasure works
correctly in that it doesn't return the bridge in B, but it returns
both `A.f` and `B.f`. This commit makes it run in an earlier phase.