Avoid quadratic complexity in iterators created by Iterator.apply
The varargs Iterator.apply now simply calls .iterator() on the varargs
parameter, instead of creating an iterator that calls the `apply` method
on the vararg. This reduces complexity, for example when passing a
List for the vararg.

For `run/t6253b.scala`, before:

    $  time ../build/quick/bin/scala Test
    166.96s user 0.61s system 100% cpu 2:46.53 total

After:

    $ time ../build/quick/bin/scala Test
    3.87s user 0.26s system 142% cpu 2.890 total

Stefan says the previous implementation was necessary to convert the
vararg collection between old and new collection types.