Prevent (LazyList|Stream).addString from forcing the stream
Originally reported as SI-8680 and fixed in Scala 2.11.

Moving partest run/t8680.scala to JUnit as LazyListTest.t8680. I changed
it back from LazyList to Stream because that is what it is supposed to
test and the laziness semantics are different. I also removed the tests
for `knownSize` of a forced stream because non-empty streams and lazy
lists never have a known size, even after they are forced. These tests
used to check `hasDefiniteSize` in 2.12 which no longer exists.