Fix #4440: Do not serialize the content of static objects
In #4450 and Scala 2.12, readResolve is used to make sure deserializing
an object returns the singleton instance of the object, but this doesn't
prevent the fields of the objects from being serialized in the first
place even though they're not used. Scala 2.13 switched to using
writeReplace to completely bypass serialization of the object in
https://github.com/scala/scala/pull/7297. This commit adapts this to
Dotty.

Co-Authored-By: Ingar Abrahamsen <ingar@abrahams1.com>
Co-Authored-By: Jason Zaugg <jzaugg@gmail.com>