Various adjustments to allow empty implicit function types
On the other hand, empty erased function types are again disabled, since
they will be compatible with no closures at all.

Note: I did not re-instantiate the tests for disallowing empty erased functions.
As is so often the case, the most trivial properties get the most tests, precisely
because they are easiest to test. But in the end, that just contributes to noise
and friction if one wants to change things. neg/i2642.scala does test the relevant bits,
that should be enough.