Select longest matching base directory in IDE
Several base source directory can match for one source file. In this
case, we select the longest match to decide what configuration to use
for a given source directory.

This case arises for instance when `sourcesInBase` is set to `true` in
sbt. In this case, the base directory of the root project (which is, in
this case, a source directory) is generally a parent of all the other
source directories.