Fix #4542: Enable Dotty LSP on `.scala` files only
Previously, the Dotty language server would try to provide services for
all `.scala` and `.sbt` files, because both are associated to the Scala
language . The association from filename to language is provided by the
extension `daltonjorge.scala`, which associates all `.scala` and `.sbt`
files to the Scala language.

This commit explicitly states that the Dotty LSP should only be used for
`.scala` files only.

Fixes #4542.