Generate JVM-compatible names for doc artifacts
We used to put the position of the doc artifact symbols
inside their names, which created a name that is not JVM-compatible.
This commit uses the fresh name creators to generate a fresh name for
every symbol introduced by the documentation.

The name is a fresh name composed of the name of the symbol that owns
the documentation, suffixed by `$doc`.