Stop rewriting withFilter to filter
This was dropped for Scala 2 in scala/scala#5252 for 2.12.0-RC1, as discussed
for instance in scala/bug#6455 and scala/bug#8339.

Beyond this code, also remove `MaybeFilter` attachment and attending code
introduced in fcea3d54dd016600f5a96cda5d03f2a5ee81e7f1.

A caveat is that, unexpectedly, this code was added in 2016-08-24, after
scala/scala#5252 was merged, as if the decision was reversed; but #1461 says:

> Also, I went through tests in pending/pos and reclassified as far as possible.

So I assume this just happened because the Scala 2 PR didn't remove our pending
test.