Deprecate override of stringPrefix
The public method, stringPrefix, for collections was initially
released in 2.13.0-M4 as deprecated and marked final.  Since
most users override this field in 2.12 and earlier, it would
produce the compiler error:

  cannot override final member

but not the deprecation warning,

  method stringPrefix is deprecated in trait Iterable (since
  2.13.0): Use className instead of stringPrefix

It would be more migration-friendly to delegate the new
method, className, to the old method, stringPrefix.  The
annotation should also have been deprecatedOverriding and not
deprecated.  That way users will see the deprecation warning:

  overriding method stringPrefix in trait Iterable is
  deprecated (since 2.13.0): Override className instead

The final keyword on stringPrefix is dropped, but the new
Iterable.className and stringPrefix are marked as protected.

Clients of a collection will no longer have access to the old
or new field. A new final field, collectionClassName, is added
as a private[scala] forwarder so that className has the proper
visibility, but the forwarder provides the exclusive access
needed by scala.runtime.ScalaRunTime#stringOf (and also a few
tests in the test suite).