Fix #1895: Use proper validity period for extension methods
When we create a symbol in a DenotationTransformer, we cannot safely
create it at phase `thisPhase.next` since this can lead to cycles. This
is why we created extension methods at phase `thisPhase`, but they're
not really valid at that point since they're only entered in their owner
once the transformer has been run. This lead to stale symbol errors
which were worked around in 51a458efeeebfeed6c357d56cf8afe5b06e86724 but
the fix is incomplete and does not work with #4604 because that PR moves
the ExtensionMethods phase around.

This commit removes all the workarounds and uses a simpler solution:
create the symbols at phase `thisPhase` but set their validity
period to match the validity of the owner transformed denotation.