Avoid allocating {Map,Set}Effect CHAMP in update/remove
These were threaded through the update/remove process to track
whether a modification was made, and if so, whether that modification
was an update to an existing map binding.

In algorithms that interact with Maps in tight loops, especially
performing no-op add/removes, these temporary object can create
heap pressure. I noticed this while tuning the DPLL implementation
in the pattern matcher.

The first concern can be handled more efficiently by checking
reference equality.

The second is rather hard to deal with without VM support for
multiple return / multi-param value classes.

The options seemed to range between:

  - Pass in a mutable result object (the status quo)
  - Encode the "replaced-ness" in the MapNode itself (which is hard
    to do without increasing its footprint or without introducing more
    subclasses and compromising bi-morphic JIT inlining)
  - redundantly re-querying the key in the input node after updated
    returns a modified node, which requires extra calls to key.hashCode
    and extra traversal through the data structure.

This commit tries a different approach. We loosen the return type
of `updated` from `MapNode` to `MapNodeSource` (of which `MapNode` is
a subclass). When an update actually does overwrite an existing
binding, we return a different subclass, `Replaced`, which wraps the
actual node and represent the extra bit of information.

This could be seen as a an encoding of `Either[MapNode, MapNode]`,
which is optimized for zero allocations in the hot paths.

In the process, I also noticed and fixed a problem with maps
failing to replace a binding if the new value was equal but
not the same reference as the previous one.

Encode replaced-ness with by returning a boxed MapNode