Update TypeMap to use assignTypes
Also, restore the invariant that the type of the TypeOf tree is the
TypeOf type itself. Here is an example showing what would go wrong
if we didn't do that. Suppose we have
  def f(x: Int) = x
  def g(x: Int) = 2 * x
  def h(f: Int => Int) = f(1)
  h(g): { 2 * 1 }
Given a type map substituting f for g in f(1), the underlying
type should be substituted to the result type of g(1), that is,
2 * 1.