Change trees in TypeOf to never loop back to their TypeOf
In the new scheme, we never touch a TypeOf tree's type and ensure this by assigning that tree NoType. This currently induces additional tree copies which could be removed at a later point.

In the process we also fixed bugs in TreeCopier, which would sometimes not go through TypeAssigner when it should have.