Add deprecated toTraversable
The toTraversable method wasn't deprecated in 2.12, so it needs to be
added back to fix clients who had used it in 2.12.  It should just be
a proxy for toIterable, because Traversable is now just a type alias
for Iterable.  After adding toTraversable back, it needs to be marked
deprecated in 2.13, so it can be removed in 2.14.

This method used to be abstract, but there's no reason it still needs
to be going forward.  Overriding toTraversable was marked deprecated
in 4f2dd84.  It should have then been marked final in 2.12, but it
probably wasn't possible since toTraversable was still overridden in
ParIterableLike and TraversableProxyLike.