Drop AppliedTypeTree
With the introduction of `Apply` nodes as types, we were facing an imbalance:
Type appications distinguished between type trees and term trees by having two
classes, `AppliedTypeTree` and `TypeApply`. Term applications did not. In the
interest of keeping the number of tree classes low, we now use `TypeApply for
both type and term trees that are applied to type tree arguments.

The only exception is in the Tasty format, where
we need to distinguish the two, because otherwise a hole followed by arguments
would be ambiguous. Distinguishing between terms and types in the Tasty serialization
format is done systematically everywhere. It has the advantage that it increases
redundancy. But for the internal tree representation it's better to have as few
classes as possible.