TreeUnpickler: Avoid forcing annotations
This avoids a cycle when unpickling scala.Predef

This change uncovered a bug when using -Ythrough-tasty: some trees were
unpickled at the wrong phase because we use
`withPhaseNoLater(ctx.picklerPhase)` in TreeUnpickler but the
TASTYCompiler previously dropped the Pickler phase, so the phase change
was a silent no-op. To avoid this issue, we change TASTYCompiler to not
drop the Pickler phase, instead we change Pickler#run to not do anything
when running with -from-tasty. We should also change how the
ctx.xxxPhase methods work to avoid this kind of silent issues.