Fix inline parameters lifted for 0 to 1
In theory the inline parameters are available in level 1,
but we can access it "unlifted" as a value in level 0.
To only pass the parameter once it is passed as a value and lifted
using the standard lifter for the values.

Unsing lifter is just an implementation detail and should not be observable
by the user. The issue was that the lifter used could be replaced or unimported
which yield unexpected error messages.