Refine handling of result types of transparent methods
1. Inline transparent methods even if result type does not match

Inlining transparent methods can improve the result type, so we should
do this anyway even if the original result type does not match.

2. Don't constrain result type of inlineable transparent methods

We need to inline the method in this case before constraining the result type.
This more systematic solution eliminates an annoying special case for typing
Apply nodes in the inliner.